Solution to day 9, but needs fixing still. The number of casts to int is horrific.

+++ src/main/java/day_07/Amplifier.java
@@ -16,9 +16,9 @@ public class Amplifier {
         return ie.isTerminated();
     }
 
-    public Integer getOutput() {
+    public Long getOutput() {
         ie.executeProgram();
-        final Queue<Integer> output = ie.getOutput();
+        final Queue<Long> output = ie.getOutput();
         return output.poll();
     }
 }
